MontyMobileKeykloak/requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.4
django-environ==0.8.1
django-model-utils==4.2.0
djangorestframework==3.13.1
drf-keycloak-auth==0.0.1.dev16
ecdsa==0.17.0
fastapi==0.75.1
fastapi_keycloak==1.0.4
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
pyasn1==0.4.8
pydantic==1.9.0
python-jose==3.3.0
python-keycloak==0.27.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
psycopg2==2.9.3
